+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Allow specifying GPG passphrase explicitly in the publish resource
+
 ## 2.3.2 (2020-11-16)
 
 - resolved cookstyle error: test/fixtures/cookbooks/aptly_test/recipes/default.rb:95:3 convention: `Style/CommentAnnotation`